Update `sample_size` method of `Dataset` trait to allow for weighted sample sizes

+++ causal-hub/src/datasets/mod.rs
@@ -29,9 +29,13 @@ pub trait Dataset {
 
     /// The sample size.
     ///
+    /// # Note
+    ///
+    /// If the dataset is weighted, this should return the sum of the weights.
+    ///
     /// # Returns
     ///
     /// The number of samples in the dataset.
     ///
-    fn sample_size(&self) -> usize;
+    fn sample_size(&self) -> f64;
 }
